+++ sc-tools/sc-server/src/sc-server-impl/sc_server_message_action.hpp
@@ -89,6 +89,7 @@ public:
     }
 
     m_server->Send(hdl, response, ScServerMessageType::text);
+    m_server->CloseConnection(hdl, websocketpp::close::status::normal, "Status checked");
   }
 
   ~ScServerMessageAction() override