+++ docs/changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `;` after `SC_ASSERT` calls in projects that use sc-machine
 
 ### Added
+- Add sc-server healthcheck
 - Add tests for sc-search agents in sc-kpm
 - Write sc-machine idea into readme
 - Add create elements by scs-helper through sc-server